Updated test harness to fuel-core 0.18.1, fuels-rs 0.42.0

+++ escrow/project/contracts/escrow-contract/tests/functions/core/take_payment.rs
@@ -40,7 +40,7 @@ mod success {
         deposit(defaults.asset_amount, &defaults.asset_id, &buyer, 0).await;
 
         provider
-            .produce_blocks(origin_block + defaults.deadline, None)
+            .produce_blocks((origin_block as u64) + defaults.deadline, None)
             .await
             .unwrap();
 
@@ -99,7 +99,7 @@ mod success {
         deposit(defaults.asset_amount, &defaults.asset_id, &buyer, 0).await;
 
         provider
-            .produce_blocks(origin_block + defaults.deadline, None)
+            .produce_blocks((origin_block as u64) + defaults.deadline, None)
             .await
             .unwrap();
 
@@ -186,7 +186,7 @@ mod success {
         .await;
 
         provider
-            .produce_blocks(origin_block + defaults.deadline, None)
+            .produce_blocks((origin_block as u64) + defaults.deadline, None)
             .await
             .unwrap();
 
@@ -321,7 +321,7 @@ mod revert {
         deposit(defaults.asset_amount, &defaults.asset_id, &buyer, 0).await;
 
         provider
-            .produce_blocks(origin_block + defaults.deadline, None)
+            .produce_blocks((origin_block as u64) + defaults.deadline, None)
             .await
             .unwrap();
 
@@ -355,7 +355,7 @@ mod revert {
         deposit(defaults.asset_amount, &defaults.asset_id, &buyer, 0).await;
 
         provider
-            .produce_blocks(origin_block + defaults.deadline, None)
+            .produce_blocks((origin_block as u64) + defaults.deadline, None)
             .await
             .unwrap();
 
@@ -387,7 +387,7 @@ mod revert {
         .await;
 
         provider
-            .produce_blocks(origin_block + defaults.deadline, None)
+            .produce_blocks((origin_block as u64) + defaults.deadline, None)
             .await
             .unwrap();
